requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
pkg-resources==0.0.0
protobuf==3.15.5
python-dateutil==2.8.1
six==1.15.0
tqdm==4.59.0
Werkzeug==1.0.1
networkx==2.5.1
